+++ service/static/service_test.go
@@ -37,25 +37,6 @@ func (cfg *testCfg) Unmarshal(out interface{}) error {
 	return j.Unmarshal([]byte(cfg.target), out)
 }
 
-func get(url string) (string, *http.Response, error) {
-	r, err := http.Get(url)
-	if err != nil {
-		return "", nil, err
-	}
-
-	b, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return "", nil, err
-	}
-
-	err = r.Body.Close()
-	if err != nil {
-		return "", nil, err
-	}
-
-	return string(b), r, err
-}
-
 func Test_Files(t *testing.T) {
 	logger, _ := test.NewNullLogger()
 	logger.SetLevel(logrus.DebugLevel)
@@ -440,6 +421,88 @@ func Test_Files_NotForbid(t *testing.T) {
 	assert.Equal(t, all("../../tests/client.php"), b)
 	assert.Equal(t, all("../../tests/client.php"), b)
 	c.Stop()
+}
+
+func TestStatic_Headers(t *testing.T) {
+	logger, _ := test.NewNullLogger()
+	logger.SetLevel(logrus.DebugLevel)
+
+	c := service.NewContainer(logger)
+	c.Register(rrhttp.ID, &rrhttp.Service{})
+	c.Register(ID, &Service{})
+
+	assert.NoError(t, c.Init(&testCfg{
+		static: `{"enable":true, "dir":"../../tests", "forbid":[], "request":{"input": "custom-header"}, "response":{"output": "output-header"}}`,
+		httpCfg: `{
+			"enable": true,
+			"address": ":8037",
+			"maxRequestSize": 1024,
+			"uploads": {
+				"dir": ` + tmpDir() + `,
+				"forbid": []
+			},
+			"workers":{
+				"command": "php ../../tests/http/client.php pid pipes",
+				"relay": "pipes",
+				"pool": {
+					"numWorkers": 1,
+					"allocateTimeout": 10000000,
+					"destroyTimeout": 10000000
+				}
+			}
+	}`}))
+
+	go func() {
+		err := c.Serve()
+		if err != nil {
+			t.Errorf("serve error: %v", err)
+		}
+	}()
+
+	time.Sleep(time.Millisecond * 500)
+
+	req, err := http.NewRequest("GET", "http://localhost:8037/client.php", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if resp.Header.Get("Output") != "output-header" {
+		t.Fatal("can't find output header in response")
+	}
+
+
+	b, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	assert.Equal(t, all("../../tests/client.php"), string(b))
+	assert.Equal(t, all("../../tests/client.php"), string(b))
+	c.Stop()
+}
+
+func get(url string) (string, *http.Response, error) {
+	r, err := http.Get(url)
+	if err != nil {
+		return "", nil, err
+	}
+
+	b, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		return "", nil, err
+	}
+
+	err = r.Body.Close()
+	if err != nil {
+		return "", nil, err
+	}
+
+	return string(b), r, err
 }
 
 func tmpDir() string {