+++ worker.go
@@ -2,14 +2,15 @@ package roadrunner
 
 import (
 	"fmt"
-	"github.com/pkg/errors"
-	"github.com/spiral/goridge/v2"
 	"os"
 	"os/exec"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
+
+	"github.com/pkg/errors"
+	"github.com/spiral/goridge/v2"
 )
 
 // Worker - supervised process with api over goridge.Relay.
@@ -249,7 +250,7 @@ func (w *Worker) execPayload(rqs *Payload) (rsp *Payload, err error) {
 	}
 
 	// add streaming support :)
-	if rsp.Body, pr, err = w.rl.Receive(); err != nil {
+	if rsp.Body, _, err = w.rl.Receive(); err != nil {
 		return nil, errors.Wrap(err, "worker error")
 	}